Fix links and buttons
