+++ site/src/components/TemplateLayout/TemplateLayout.tsx
@@ -1,22 +1,8 @@
-import Button from "@material-ui/core/Button"
-import Link from "@material-ui/core/Link"
 import { makeStyles } from "@material-ui/core/styles"
-import AddCircleOutline from "@material-ui/icons/AddCircleOutline"
-import SettingsOutlined from "@material-ui/icons/SettingsOutlined"
 import { useMachine } from "@xstate/react"
-import {
-  PageHeader,
-  PageHeaderSubtitle,
-  PageHeaderTitle,
-} from "components/PageHeader/PageHeader"
 import { useOrganizationId } from "hooks/useOrganizationId"
 import { createContext, FC, Suspense, useContext } from "react"
-import {
-  Link as RouterLink,
-  NavLink,
-  Outlet,
-  useParams,
-} from "react-router-dom"
+import { NavLink, Outlet, useParams } from "react-router-dom"
 import { combineClasses } from "util/combineClasses"
 import {
   TemplateContext,
@@ -27,14 +13,7 @@ import { Stack } from "components/Stack/Stack"
 import { Permissions } from "xServices/auth/authXService"
 import { Loader } from "components/Loader/Loader"
 import { usePermissions } from "hooks/usePermissions"
-import { Avatar } from "components/Avatar/Avatar"
-
-const Language = {
-  settingsButton: "Settings",
-  editButton: "Edit",
-  createButton: "Create workspace",
-  noDescription: "",
-}
+import { TemplatePageHeader } from "./TemplatePageHeader"
 
 const useTemplateName = () => {
   const { template } = useParams()
@@ -65,35 +44,6 @@ export const useTemplateLayoutContext = (): TemplateLayoutContextValue => {
   return context
 }
 
-const TemplateSettingsButton: FC<{ templateName: string }> = ({
-  templateName,
-}) => (
-  <Link
-    underline="none"
-    component={RouterLink}
-    to={`/templates/${templateName}/settings`}
-  >
-    <Button variant="outlined" startIcon={<SettingsOutlined />}>
-      {Language.settingsButton}
-    </Button>
-  </Link>
-)
-
-const CreateWorkspaceButton: FC<{
-  templateName: string
-  className?: string
-}> = ({ templateName, className }) => (
-  <Link
-    underline="none"
-    component={RouterLink}
-    to={`/templates/${templateName}/workspace`}
-  >
-    <Button className={className ?? ""} startIcon={<AddCircleOutline />}>
-      {Language.createButton}
-    </Button>
-  </Link>
-)
-
 export const TemplateLayout: FC<{ children?: JSX.Element }> = ({
   children = <Outlet />,
 }) => {
@@ -108,58 +58,17 @@ export const TemplateLayout: FC<{ children?: JSX.Element }> = ({
   })
   const { template, permissions: templatePermissions } = templateState.context
   const permissions = usePermissions()
-  const hasIcon = template && template.icon && template.icon !== ""
 
   if (!template || !templatePermissions) {
     return <Loader />
   }
 
-  const generatePageHeaderActions = (): JSX.Element[] => {
-    const pageActions: JSX.Element[] = []
-
-    if (templatePermissions?.canUpdateTemplate) {
-      pageActions.push(<TemplateSettingsButton templateName={template.name} />)
-    }
-
-    pageActions.push(<CreateWorkspaceButton templateName={template.name} />)
-
-    return pageActions
-  }
-
   return (
     <>
-      <Margins>
-        <PageHeader
-          actions={
-            <>
-              {generatePageHeaderActions().map((action, i) => (
-                <div key={i}>{action}</div>
-              ))}
-            </>
-          }
-        >
-          <Stack direction="row" spacing={3} className={styles.pageTitle}>
-            {hasIcon ? (
-              <Avatar size="xl" src={template.icon} variant="square" fitImage />
-            ) : (
-              <Avatar size="xl">{template.name}</Avatar>
-            )}
-
-            <div>
-              <PageHeaderTitle>
-                {template.display_name.length > 0
-                  ? template.display_name
-                  : template.name}
-              </PageHeaderTitle>
-              <PageHeaderSubtitle condensed>
-                {template.description === ""
-                  ? Language.noDescription
-                  : template.description}
-              </PageHeaderSubtitle>
-            </div>
-          </Stack>
-        </PageHeader>
-      </Margins>
+      <TemplatePageHeader
+        template={template}
+        permissions={templatePermissions}
+      />
 
       <div className={styles.tabs}>
         <Margins>
@@ -204,17 +113,6 @@ export const TemplateLayout: FC<{ children?: JSX.Element }> = ({
 
 export const useStyles = makeStyles((theme) => {
   return {
-    pageTitle: {
-      alignItems: "center",
-    },
-    iconWrapper: {
-      width: theme.spacing(6),
-      height: theme.spacing(6),
-      "& img": {
-        width: "100%",
-      },
-    },
-
     tabs: {
       borderBottom: `1px solid ${theme.palette.divider}`,
       marginBottom: theme.spacing(5),
